Add files via upload
A small pizza company requires an analyst to retrieve data to be used to help make strategic business decisions.
This Word document captures the select statements that can be used to retrieve records for the pizza company. Some
of the records/data were retrieved using aggregates or math calculations that return single values based on column data.